Move manager from editor utils to switcher plugin
