+++ spyder/plugins/switcher/manager.py
